+++ lib/utils.ex
@@ -8,6 +8,8 @@ defmodule Unicode.Utils do
   @scripts_path Path.join(Unicode.data_dir(), "scripts.txt")
   def scripts do
     parse_file(@scripts_path)
+    |> downcase_keys()
+    |> atomize_keys()
   end
 
   @doc """
@@ -17,16 +19,8 @@ defmodule Unicode.Utils do
   @blocks_path Path.join(Unicode.data_dir(), "blocks.txt")
   def blocks do
     parse_file(@blocks_path)
-    |> Enum.map(fn {k, v} ->
-      new_key =
-        k
-        |> String.replace(" ", "_")
-        |> String.replace("-", "_")
-        |> String.to_atom()
-
-      {new_key, v}
-    end)
-    |> Map.new()
+    |> downcase_keys
+    |> atomize_keys()
   end
 
   @doc """
@@ -37,7 +31,7 @@ defmodule Unicode.Utils do
   def combining_classes do
     parse_file(@combining_class_path)
     |> Enum.map(fn {k, v} -> {String.to_integer(k), v} end)
-    |> Enum.into(%{})
+    |> Map.new
   end
 
   @doc """
@@ -47,49 +41,18 @@ defmodule Unicode.Utils do
   @categories_path Path.join(Unicode.data_dir(), "categories.txt")
   def categories do
     parse_file(@categories_path)
-    |> Enum.map(fn {k, v} -> {titlecase(k), v} end)
-    |> atomize_keys
-  end
-
-  @test_path Path.join(Unicode.data_dir(), "test.txt")
-  def test do
-    parse_file(@test_path)
-    |> Enum.map(fn {k, v} -> {titlecase(k), v} end)
-    |> atomize_keys
+    |> atomize_keys()
   end
 
   @doc """
-  Returns a map of the Unicode codepoints with the emoji type name
+  Returns a map of the Unicode codepoints with the `derived property` name
   as the key and a list of codepoint ranges as the values.
   """
-  @emoji_path Path.join(Unicode.data_dir(), "emoji.txt")
-  def emoji do
-    parse_file(@emoji_path)
-    |> Enum.map(fn {k, v} -> {String.downcase(k), v} end)
-    |> atomize_keys
-  end
-
-  @doc """
-  Returns a map of the property value aliases.
-  """
-  @property_value_alias_path Path.join(Unicode.data_dir(), "property_value_alias.txt")
-  def property_value_alias do
-    Enum.reduce(File.stream!(@property_value_alias_path), [], fn line, acc ->
-      case line do
-        <<"#", _rest::bitstring>> ->
-          acc
-
-        <<"\n", _rest::bitstring>> ->
-          acc
-
-      data ->
-        [data
-        |> String.replace(~r/ *#.*/, "")
-        |> String.split(";")
-        |> Enum.map(&String.trim/1) | acc]
-      end
-    end)
-    |> Enum.group_by(&hd/1, &tl/1)
+  @derived_properties_path Path.join(Unicode.data_dir(), "derived_properties.txt")
+  def derived_properties do
+    parse_file(@derived_properties_path)
+    |> downcase_keys
+    |> atomize_keys()
   end
 
   @doc """
@@ -99,7 +62,66 @@ defmodule Unicode.Utils do
   @properties_path Path.join(Unicode.data_dir(), "properties.txt")
   def properties do
     parse_file(@properties_path)
-    |> atomize_keys
+    |> downcase_keys()
+    |> atomize_keys()
+  end
+
+  @doc """
+  Returns a map of the Unicode codepoints with the emoji type name
+  as the key and a list of codepoint ranges as the values.
+  """
+  @emoji_path Path.join(Unicode.data_dir(), "emoji.txt")
+  def emoji do
+    parse_file(@emoji_path)
+    |> downcase_keys
+    |> atomize_keys()
+  end
+
+  @doc """
+  Returns a map of the property value aliases.
+  """
+  @property_alias_path Path.join(Unicode.data_dir(), "property_alias.txt")
+  def property_alias do
+    parse_alias_file(@property_alias_path)
+    |> Enum.flat_map(fn
+      [alias1, code] ->
+        [{alias1, code}]
+
+      [alias1, code, alias2] ->
+        [{alias1, code}, {alias2, code}]
+
+      [alias1, code, alias2, alias3] ->
+        [{alias1, code}, {alias2, code}, {alias3, code}]
+
+      [_alias1, _code, _alias2, _alias3, _alias4] ->
+        []
+    end)
+    |> Map.new()
+  end
+
+  @doc """
+  Returns a map of the property value aliases.
+  """
+  @property_value_alias_path Path.join(Unicode.data_dir(), "property_value_alias.txt")
+  def property_value_alias do
+    parse_alias_file(@property_value_alias_path)
+    |> Enum.group_by(&hd/1, &tl/1)
+    |> Enum.map(fn {category, aliases} -> {category, map_from_aliases(aliases)} end)
+    |> Map.new()
+  end
+
+  defp map_from_aliases(aliases) do
+    Enum.flat_map(aliases, fn
+      [code, alias1] ->
+        [{alias1, code}]
+
+      [code, alias1, alias2] ->
+        [{alias1, code}, {alias2, code}]
+
+      [code, alias1, alias2, alias3] ->
+        [{alias1, code}, {alias2, code}, {alias3, code}]
+    end)
+    |> Map.new()
   end
 
   @doc false
@@ -150,7 +172,7 @@ defmodule Unicode.Utils do
       end
     end)
     |> Enum.map(fn {key, ranges} ->
-      {String.downcase(key), Enum.reverse(ranges)}
+      {key, Enum.reverse(ranges)}
     end)
     |> Enum.into(%{})
   end
@@ -173,6 +195,67 @@ defmodule Unicode.Utils do
       codepoints ->
         Enum.map(codepoints, &String.to_integer(&1, 16))
     end
+  end
+
+  @doc false
+  def parse_alias_file(path) do
+    Enum.reduce(File.stream!(path), [], fn line, acc ->
+      case line do
+        <<"#", _rest::bitstring>> ->
+          acc
+
+        <<"\n", _rest::bitstring>> ->
+          acc
+
+        data ->
+          [
+            data
+            |> String.replace(~r/ *#.*/, "")
+            |> String.split(";")
+            |> Enum.map(fn n -> String.trim(n) |> String.downcase() end)
+            | acc
+          ]
+      end
+    end)
+  end
+
+  # Take the atom values of the map
+  # and add a string version as an alias
+  def add_canonical_alias(map) do
+    map
+    |> Enum.map(fn {_k, v} -> {downcase_and_remove_whitespace(v), v} end)
+    |> Map.new
+    |> Map.merge(map)
+  end
+
+  def downcase_keys_and_remove_whitespace(map) when is_map(map) do
+    Enum.map(map, fn {k, v} -> {downcase_and_remove_whitespace(k), v} end)
+    |> Map.new()
+  end
+
+  @match [" ", "-", "_"]
+  def downcase_and_remove_whitespace(string) when is_binary(string) do
+    string
+    |> String.trim()
+    |> String.downcase()
+    |> String.replace(@match, "")
+  end
+
+  def downcase_and_remove_whitespace(atom) when is_atom(atom) do
+    atom
+    |> Atom.to_string()
+    |> downcase_and_remove_whitespace()
+  end
+
+  def downcase_and_remove_whitespace(integer) when is_integer(integer) do
+    integer
+    |> Integer.to_string()
+  end
+
+  def conform_key(string) do
+    string
+    |> String.replace(" ", "_")
+    |> String.replace("-", "_")
   end
 
   @doc """
@@ -235,14 +318,32 @@ defmodule Unicode.Utils do
   end
 
   @doc false
-  defp titlecase(string) do
-    String.capitalize(string)
+  def capitalize_keys(map) do
+    Enum.map(map, fn {k, v} -> {String.capitalize(k), v} end)
+    |> Map.new()
+  end
+
+  @doc false
+  def downcase_keys(map) do
+    Enum.map(map, fn {k, v} -> {String.downcase(k), v} end)
+    |> Map.new()
   end
 
   @doc false
   def atomize_keys(map) do
-    Enum.map(map, fn {k, v} -> {String.to_atom(k), v} end)
-    |> Enum.into(%{})
+    Enum.map(map, fn {k, v} -> {String.to_atom(conform_key(k)), v} end)
+    |> Map.new()
+  end
+
+  @doc false
+  def capitalize_values(map) do
+    Enum.map(map, fn {k, v} -> {k, String.capitalize(v)} end)
+    |> Map.new()
+  end
+
+  def atomize_values(map) do
+    Enum.map(map, fn {k, v} -> {k, String.to_atom(conform_key(v))} end)
+    |> Map.new()
   end
 
   @doc false
@@ -280,5 +381,11 @@ defmodule Unicode.Utils do
       {first, last}, acc ->
         Enum.map(last..first, & &1) ++ acc
     end)
+  end
+
+  @doc false
+  def reverse_map(map) do
+    Enum.map(map, fn {k, v} -> {v, k} end)
+    |> Map.new()
   end
 end